add tests using UnionML tasks in Flyte workflows
