+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@ repos:
   hooks:
   - id: shellcheck
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.942
+  rev: v0.971
   hooks:
     - id: mypy
       entry: mypy unionml tests
